Now using the useSession hook to fix bug. Added more logic to options.js jwt callback
